MyoPS_Segmentation - 副本/requirements.txt
tensorflow
nibabel
numpy
scikit-image
scipy
imgaug
pyyaml
matplotlib
tqdm